+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 - Compasses are no longer too large when held (#61)
+- You can now fill Glass Bottles from Pitcher Plants (#82)
 
 # 1.11.0